feat: implement useSupportedFields hook for managing identity field retrieval
